rename : service 패키지 삭제, service 파일 application 패키지로 이동
